+++ app/javascript/components/addPage/FoundInDatabase.js
@@ -1,12 +1,12 @@
-import React from 'react'
+import React from 'react';
 
-const AlreadyInDB = ({ title }) => {
+const FoundInDatabase = ({ title }) => {
   return(
     <div className="in-database-message">
       <p>{`We found ${title} in our collection, so you shouldn't need a web search to add that one. `}
       If it didn't appear, double check that you haven't already added it.</p>
     </div>
-  )
-}
+  );
+};
 
-export default AlreadyInDB
+export default FoundInDatabase;
