+++ app/javascript/components/addPage/SearchPrompt.js
