+++ app/javascript/components/addPage/SelectFormType.js
