+++ app/javascript/components/homePage/AddMediaPrompt.js
