+++ app/javascript/components/shared/FlashNotice.js
