+++ app/javascript/components/shared/RatingInput.js
