+++ app/javascript/components/showPage/MediaDetailsDisplay.js
